Functional Mod Manager
The manager can actually activate and deactivate mods now. The manager
will not remove directories (as a security precaution until I make the
removal safer). The manager still needs to pack its .dll, mod detection
needs to be activated (hashing in general needs to be added), and I need
to add a more tasteful easter egg.